make icons responsive
